move python files to sub-directory #5
